Add automatic discovery for location of custom module
Instead of letting the user specify where to find custom modules, use
`sys.path` to auto-search for potential matches. This simplifies the
process and reduces the burden on the user. For each possible import
base, we look for a potential match and break right away if one is not
found.