Unquote the URL instead of trying to hardcode for localhost cases
If we're using MDB in a backend that sends back IPs and ports, then we
need to be able to understand them to send back. Right now we do this in
a hard-coded way for localhost, but we need a more robust solution.

This change limits slightly the valid names, but that shouldn't be an
issue.